[Fix for ticket #2] Cloak status is visible on the in-flight map again.
I probably introduced this issue in r41 when I improved the handling of the 'Current Location' and 'Currently Viewing' displays. The cloak status would only be shown for 'Currently Viewing' which only happened on the map when in base.

This change splits it up some more, creating separate images for !LocationInCockpitImage and !LocationOnMapImage. The latter contains the cloak status again, the former does not (because all the HUDs feature their own cloak status -- no need to duplicate that when in the cockpit).

For this to work, the definition of !CloakOn had to be moved up, away from its corresponding definitions. Unfortunately and could be refactored, but I guess it doesn't matter much at the current state of the `dialog.mdl`.