[Partial fix for issue #26] Created a very rough proof-of-concept for a custom action for Modulus to configure CortUI with.
This custom action pretty much just replicates the configuration screen that is already included in the CortUI installer. It presents a configuration dialog to the user and then writes the cortui_settings.mdl accordingly.

The uninstaller currently just deletes the cortui_settings.mdl.

It has not been successfully tested with Modulus yet.

The CortUI version and the cortui_settings.mdl template are currently hardcoded into the custom action. This should be pulled out in the future for easier distribution package creation.